# jaxrs-spec-generated/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Jouko API spec",
    "version" : "0.0.4",
    "title" : "Jouko API"
  },
  "basePath" : "/v1",
  "schemes" : [ "https" ],
  "consumes" : [ "application/json;charset=utf-8" ],
  "produces" : [ "application/json;charset=utf-8" ],
  "security" : [ {
    "bearer" : [ ]
  } ],
  "paths" : {
...
    "/users/{userId}/powerMeasurements" : {
      "get" : {
        "tags" : [ "Measurements" ],
        "summary" : "List all measurements",
        "operationId" : "listAllMeasurements",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "description" : "The id of the user",
          "required" : true,
          "type" : "integer",
          "format" : "int6"
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/PowerMeasurement"
              }
            }
          },
          "401" : {
            "description" : "Unauthorized",
            "schema" : {
              "$ref" : "#/definitions/Unauthorized"
            }
          },
          "500" : {
            "description" : "Internal server error",
            "schema" : {
              "$ref" : "#/definitions/InternalServerError"
            }
          }
        }
      }
    },
...
    "/admin/users/{userId}" : {
      "get" : {
        "tags" : [ "Users" ],
        "summary" : "Retreive user",
        "operationId" : "retrieveUser",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "description" : "The id of the user",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "$ref" : "#/definitions/User"
            }
          },
          "401" : {
            "description" : "Unauthorized",
            "schema" : {
              "$ref" : "#/definitions/Unauthorized"
            }
          },
          "500" : {
            "description" : "Internal server error",
            "schema" : {
              "$ref" : "#/definitions/InternalServerError"
            }
          }
        }
      },
      "put" : {
        "tags" : [ "Users" ],
        "summary" : "Update user",
        "operationId" : "updateUser",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "description" : "The id of the user",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "in" : "body",
          "name" : "body",
          "description" : "The user to be updated",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/User"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "$ref" : "#/definitions/User"
            }
          },
          "401" : {
            "description" : "Unauthorized",
            "schema" : {
              "$ref" : "#/definitions/Unauthorized"
            }
          },
          "500" : {
            "description" : "Internal server error",
            "schema" : {
              "$ref" : "#/definitions/InternalServerError"
            }
          }
        }
      }
    },
    "/keycloakUsers/{keycloakId}" : {
      "get" : {
        "tags" : [ "Users" ],
        "summary" : "Get an user by keycloak id",
        "operationId" : "getUserByKeycloakId",
        "parameters" : [ {
          "name" : "keycloakId",
          "in" : "path",
          "description" : "The keycloak id of the user to look up",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "$ref" : "#/definitions/User"
            }
          },
          "401" : {
            "description" : "Unauthorized",
            "schema" : {
              "$ref" : "#/definitions/Unauthorized"
            }
          },
          "500" : {
            "description" : "Internal server error",
            "schema" : {
              "$ref" : "#/definitions/InternalServerError"
            }
          }
        }
      }
    },
    "/users/{userId}/interruptions" : {
      "get" : {
        "tags" : [ "Interruptions" ],
        "summary" : "List interruptions",
        "operationId" : "listInterruptions",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "description" : "The user whose interruptions we list",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "name" : "fromTime",
          "in" : "query",
          "description" : "List interruptions that start after this point of time, inclusive",
          "required" : true,
          "type" : "string",
          "format" : "date-time"
        }, {
          "name" : "toTime",
          "in" : "query",
          "description" : "List interruptions that end before this point of time, exclusive",
          "required" : true,
          "type" : "string",
          "format" : "date-time"
        }, {
          "name" : "deviceId",
          "in" : "query",
          "description" : "The device id whose interruptions we list, list all if omitted",
          "required" : false,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Interruption"
              }
            }
          },
          "401" : {
            "description" : "Unauthorized",
            "schema" : {
              "$ref" : "#/definitions/Unauthorized"
            }
          },
          "500" : {
            "description" : "Internal server error",
            "schema" : {
              "$ref" : "#/definitions/InternalServerError"
            }
          }
        }
      }
    },
    "/users/{userId}/interruptions/{interruptionId}/cancelled" : {
      "put" : {
        "tags" : [ "Interruptions" ],
        "summary" : "Set interruption's cancellation status",
        "operationId" : "setInterruptionCancelled",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "description" : "The user who owns the interruption",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "name" : "interruptionId",
          "in" : "path",
          "description" : "The id of the interruption to set cancellation status of",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "in" : "body",
          "name" : "body",
          "description" : "The cancellation status",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/InterruptionCancellation"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "$ref" : "#/definitions/InterruptionCancellation"
            }
          },
          "401" : {
            "description" : "Unauthorized",
            "schema" : {
              "$ref" : "#/definitions/Unauthorized"
            }
          },
          "500" : {
            "description" : "Internal server error",
            "schema" : {
              "$ref" : "#/definitions/InternalServerError"
            }
          }
        }
      }
    },
    "/users/{userId}/devices" : {
      "get" : {
        "tags" : [ "Devices" ],
        "summary" : "List devices",
        "operationId" : "listDevices",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "description" : "The user who owns the interruption",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "name" : "firstResult",
          "in" : "query",
          "description" : "The offset of the first result",
          "required" : true,
          "type" : "integer"
        }, {
          "name" : "maxResults",
          "in" : "query",
          "description" : "The maximum number of results",
          "required" : true,
          "type" : "integer"
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Device"
              }
            }
          },
          "401" : {
            "description" : "Unauthorized",
            "schema" : {
              "$ref" : "#/definitions/Unauthorized"
            }
          },
          "500" : {
            "description" : "Internal server error",
            "schema" : {
              "$ref" : "#/definitions/InternalServerError"
            }
          }
        }
      }
    },
    "/users/{userId}/devices/{deviceId}/powerConsumption" : {
      "get" : {
        "tags" : [ "Devices" ],
        "summary" : "Get the power consumption of the given device in a time period",
        "operationId" : "getPowerConsumption",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "description" : "The user that owns the device",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "name" : "deviceId",
          "in" : "path",
          "description" : "The device whose power consumption we measure",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "name" : "fromTime",
          "in" : "query",
          "description" : "The start of the time period that we measure the consumption in, inclusive",
          "required" : true,
          "type" : "string",
          "format" : "date-time"
        }, {
          "name" : "toTime",
          "in" : "query",
          "description" : "The end of the time period that we measure the consumption in, exclusive",
          "required" : true,
          "type" : "string",
          "format" : "date-time"
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "$ref" : "#/definitions/DevicePowerConsumption"
            }
          },
          "401" : {
            "description" : "Unauthorized",
            "schema" : {
              "$ref" : "#/definitions/Unauthorized"
            }
          },
          "500" : {
            "description" : "Internal server error",
            "schema" : {
              "$ref" : "#/definitions/InternalServerError"
            }
          }
        }
      }
    }
  },
  "securityDefinitions" : {
    "bearer" : {
      "type" : "apiKey",
      "name" : "Authorization",
      "in" : "header"
    }
  },
  "definitions" : {
    "Unauthorized" : {
      "type" : "object",
      "required" : [ "message" ],
      "properties" : {
        "message" : {
          "type" : "string",
          "description" : "The error message"
        }
      }
    },
    "InternalServerError" : {
      "type" : "object",
      "required" : [ "message" ],
      "properties" : {
        "message" : {
          "type" : "string",
          "description" : "The error message"
        }
      }
    },
    "User" : {
      "type" : "object",
      "required" : [ "id", "keycloakId" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "User id"
        },
        "keycloakId" : {
          "type" : "string",
          "format" : "uuid",
          "description" : "The keycloak id of the user"
        }
      }
    },
...
    "Interruption" : {
      "type" : "object",
      "required" : [ "cancelled", "deviceId", "endTime", "id", "startTime" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "Interruption id"
        },
        "startTime" : {
          "type" : "string",
          "format" : "date-time",
          "description" : "Start time of the interruption, inclusive"
        },
        "endTime" : {
          "type" : "string",
          "format" : "date-time",
          "description" : "End time of the interruption, exclusive"
        },
        "deviceId" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "The ID of the interrupted device"
        },
        "cancelled" : {
          "type" : "boolean",
          "description" : "True if the interruption has been cancelled"
        },
        "cancellationTime" : {
          "type" : "string",
          "format" : "date-time",
          "description" : "The time when the interruption was cancelled"
        }
      }
    },
    "InterruptionCancellation" : {
      "type" : "object",
      "required" : [ "cancelled" ],
      "properties" : {
        "cancelled" : {
          "type" : "boolean",
          "description" : "True if the interruption has been cancelled"
        }
      }
    },
    "PowerMeasurement" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "Measurement id"
        },
        "startTime" : {
          "type" : "string",
          "format" : "date-time",
          "description" : "Start time of the measurement"
        },
        "endTime" : {
          "type" : "string",
          "format" : "date-time",
          "description" : "End time of the measurement"
        },
        "measurementType" : {
          "type" : "string",
          "description" : "Type of measurement",
          "enum" : [ "AVERAGE" ]
        },
        "measurementValue" : {
          "type" : "number",
          "format" : "double"
        },
        "device_id" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "Device id"
        },
        "phaseNumber" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "Phasenumber"
        }
      }
    },
    "Device" : {
      "type" : "object",
      "required" : [ "id", "name" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "Device id"
        },
        "name" : {
          "type" : "string",
          "description" : "The name of the device"
        },
        "userId" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "Device user id"
        },
        "controllerId" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "Device controller id"
        }
      }
    },
...
    "DevicePowerConsumption" : {
      "type" : "object",
      "required" : [ "averageConsumptionInWatts" ],
      "properties" : {
        "averageConsumptionInWatts" : {
          "type" : "number",
          "format" : "double",
          "description" : "The power consumption in the given time span, in watts"
        }
      }
    }
  }
}